mkfs: fix memory corruption bug
- Prevented a memory corruption bug. This wasn't the only thing changed, but I think it's really interesting so here's an explanation to why this happened.

The maximum file size of the xv6 filesystem is 140 blocks, also 70 kilobytes. libc.a, which is 74 kilobytes and 148 blocks, exceeds the 70 kilobytes limit.

MIT never implemented a check to reject files if they exceed 70 kilobytes.

This should be standard practice as this is programmed in C because if you don't implement a check for something like this, you'll cause memory corruption.

The iappend function is called which allocates the blocks for libc.a. Blocks 0-11 are filled which is fine. From 12-139, it uses the indirect blocks (0-127) which is also fine.

Block 139 (indirect block 127) is the last valid element it can use. When we hit block 140 though... we move to indirect 128... which isn't valid memory and now we've buffer overflowed and are using garbage on the stack!!!

Since we are now using the stack for indirect 128, it's pretty much random as it's whatever next to the indirect array in memory. This is passed to rsect, which gives "failed to read disk block: Success" errors.

rsect has to seek to that garbage, which is way past the 1000K fs.img. the read call returns 0 here as it's EOF.  the read check is then true, which throws up that disk block failure message with success in it (which threw us off completely during initial investigations).

The garbage is then passed to wsect, which seeks just like rseek. lseek succeeds since it can read past the end of files and the write call succeeds and instantly writes a mostly zeroes 3.7GB fs.img (at least it's that big for me).

Following this, iappend was mostly rewritten. the limit was extended from 70 kilobytes to 1 gigabyte and a file limit check was put in place. How was this not caught sooner? I have no clue.

- Implemented a 3-level indirect block system.

- FSSIZE is now 10000 blocks.

- Remove the include/kernel/fs.h requirement from mkfs, I'm not sure why that was there.

- Sync tool and kernel includes.

- This is my best commit yet, I think.